+++ src/pages/constants/Clone.tsx
@@ -15,7 +15,7 @@ const section = SECTIONS_MAP.constant
 export const Component = () => {
     const queryFn = useBoundResourceQueryFn()
     const [searchParams] = useSearchParams()
-    const duplicatedModelId = searchParams.get('duplicatedId') as string
+    const duplicatedModelId = searchParams.get('clonedId') as string
 
     const query = {
         resource: 'constants',
