+++ src/pages/indicatorTypes/Clone.tsx
@@ -18,7 +18,7 @@ type IndicatorTypesDuplicateFormValues = IndicatorTypesFormValues
 export const Component = () => {
     const queryFn = useBoundResourceQueryFn()
     const [searchParams] = useSearchParams()
-    const duplicatedModelId = searchParams.get('duplicatedId') as string
+    const duplicatedModelId = searchParams.get('clonedId') as string
 
     const query = {
         resource: 'indicatorTypes',
