+++ src/pages/legendSets/Clone.tsx
@@ -31,7 +31,7 @@ const section = SECTIONS_MAP.legendSet
 export const Component = () => {
     const queryFn = useBoundResourceQueryFn()
     const [searchParams] = useSearchParams()
-    const duplicatedModelId = searchParams.get('duplicatedId') as string
+    const duplicatedModelId = searchParams.get('clonedId') as string
 
     const query = {
         resource: 'legendSets',
