+++ src/pages/organisationUnits/list/OrganisationUnitListActions.tsx
@@ -84,8 +84,8 @@ const OrganisationUnitActionMore = ({
     )
 
     const handleDuplicateClick = useLinkClickHandler({
-        pathname: 'duplicate',
-        search: `?duplicatedId=${model.id}`,
+        pathname: 'clone',
+        search: `?clonedId=${model.id}`,
     })
 
     const handleAddChildClick = useLinkClickHandler(
@@ -127,7 +127,7 @@ const OrganisationUnitActionMore = ({
                         />
                         <MenuItem
                             dense
-                            label={i18n.t('Duplicate')}
+                            label={i18n.t('Clone')}
                             icon={<IconDuplicate16 />}
                             onClick={(_, e) => {
                                 handleDuplicateClick(e)
