+++ src/pages/programIndicators/Clone.tsx
@@ -23,7 +23,7 @@ const section = SECTIONS_MAP.programIndicator
 export const Component = () => {
     const queryFn = useBoundResourceQueryFn()
     const [searchParams] = useSearchParams()
-    const duplicatedModelId = searchParams.get('duplicatedId') as string
+    const duplicatedModelId = searchParams.get('clonedId') as string
 
     const programIndicators = useQuery({
         queryFn: queryFn<ProgramIndicatorValues>,
