+++ src/pages/validationRules/Clone.tsx
@@ -51,7 +51,7 @@ const section = SECTIONS_MAP.validationRule
 export const Component = () => {
     const queryFn = useBoundResourceQueryFn()
     const [searchParams] = useSearchParams()
-    const duplicatedModelId = searchParams.get('duplicatedId') as string
+    const duplicatedModelId = searchParams.get('clonedId') as string
 
     const query = {
         resource: 'validationRules',
